+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # HEAD
 
+* Use Psych to generate the yaml if it's available
 * Report when the directory to write the yml to doesn't exist (Thanks Robin Mehner)
 
 # Version 0.3.2